Update test globs (#808)
