+++ packages/create-yoshi-app/templates/business-manager-module-typescript/__tests__/e2e-setup.ts
