+++ packages/create-yoshi-app/templates/business-manager-module-typescript/__tests__/test-common.ts
