+++ packages/create-yoshi-app/templates/business-manager-module-typescript/environment.ts
