+++ packages/create-yoshi-app/templates/business-manager-module-typescript/src/components/__mocks__/i18n.ts
