+++ packages/create-yoshi-app/templates/business-manager-module/__tests__/dev/server.js
@@ -1,4 +1,4 @@
-const testKitEnv = require('../environment')
+const testKitEnv = require('../../environment')
   .environment()
   .then(env => {
     env.start();
@@ -8,13 +8,12 @@ const testKitEnv = require('../environment')
 // We need to stop the testkit explicitly, since it's running in a different process
 const stopTestKit = () => testKitEnv.then(tk => tk.stop());
 
-const signals: ('SIGINT' | 'SIGUSR1' | 'SIGUSR2')[] = [
+const stopEvents = [
   'SIGINT',
   'SIGUSR1',
   'SIGUSR2',
+  'uncaughtException',
+  'exit',
 ];
 
-signals.forEach(ev => process.on(ev, stopTestKit));
-
-process.on('uncaughtException', stopTestKit);
-process.on('exit', stopTestKit);
+stopEvents.forEach(ev => process.on(ev, stopTestKit));
