+++ packages/create-yoshi-app/templates/business-manager-module/__tests__/e2e-setup.js
