+++ packages/create-yoshi-app/templates/business-manager-module/__tests__/e2e/app.e2e.js
