+++ packages/create-yoshi-app/templates/business-manager-module/__tests__/spec-setup.js
