+++ packages/create-yoshi-app/templates/business-manager-module/__tests__/test-common.js
