+++ packages/create-yoshi-app/templates/business-manager-module/environment.js
