+++ packages/create-yoshi-app/templates/business-manager-module/jest-yoshi.config.js
@@ -1,4 +1,4 @@
-const { environment } = require('./test/environment');
+const { environment } = require('./environment');
 
 // The purpose of this file is to start your server and possibly additional servers
 // like RPC/Petri servers.
