+++ packages/create-yoshi-app/templates/business-manager-module/src/components/__mocks__/i18n.js
