+++ packages/create-yoshi-app/templates/client-typescript/__tests__/e2e-setup.ts
