+++ packages/create-yoshi-app/templates/client-typescript/__tests__/simple.e2e.ts
