+++ packages/create-yoshi-app/templates/client-typescript/src/components/__mocks__/i18n.ts
