+++ packages/create-yoshi-app/templates/client/__tests__/e2e-setup.js
