+++ packages/create-yoshi-app/templates/client/__tests__/simple.e2e.js
