+++ packages/create-yoshi-app/templates/client/src/components/__mocks__/i18n.js
