+++ packages/create-yoshi-app/templates/fullstack-typescript/__tests__/e2e-setup.ts
