+++ packages/create-yoshi-app/templates/fullstack-typescript/__tests__/server.e2e.ts
