+++ packages/create-yoshi-app/templates/fullstack-typescript/__tests__/simple.e2e.ts
