+++ packages/create-yoshi-app/templates/fullstack-typescript/environment.js
