+++ packages/create-yoshi-app/templates/fullstack-typescript/src/components/__mocks__/i18n.ts
