+++ packages/create-yoshi-app/templates/fullstack/__tests__/e2e-setup.js
