+++ packages/create-yoshi-app/templates/fullstack/__tests__/server.e2e.js
