+++ packages/create-yoshi-app/templates/fullstack/__tests__/simple.e2e.js
