+++ packages/create-yoshi-app/templates/fullstack/environment.js
