+++ packages/create-yoshi-app/templates/fullstack/jest-yoshi.config.js
@@ -1,4 +1,4 @@
-const { emitConfigs, bootstrapServer } = require('./test/environment');
+const { emitConfigs, bootstrapServer } = require('./environment');
 
 // The purpose of this file is to start your server and possibly additional servers
 // like RPC/Petri servers.
