+++ packages/create-yoshi-app/templates/fullstack/src/components/__mocks__/i18n.js
