+++ packages/create-yoshi-app/templates/server-typescript/{%packagejson%}
@@ -11,7 +11,7 @@
     "start": "yoshi start",
     "precommit": "lint-staged",
     "pretest": "yoshi build",
-    "test": "yoshi test",
+    "test": "yoshi test --mocha",
     "posttest": "yoshi lint",
     "release": "yoshi release"
   },
