+++ packages/jest-yoshi-preset/jest-preset.js
@@ -1,7 +1,7 @@
 const fs = require('fs');
 const globby = require('globby');
 const { envs } = require('./constants');
-const globs = require('../yoshi-config/globs');
+const globs = require('yoshi-config/globs');
 
 const modulePathIgnorePatterns = ['<rootDir>/dist/', '<rootDir>/target/'];
 module.exports = {
@@ -12,25 +12,15 @@ module.exports = {
   projects: [
     ...[
       {
-        displayName: 'component',
+        displayName: 'spec',
         testEnvironment: 'jsdom',
         testURL: 'http://localhost',
-        testMatch: ['<rootDir>/src/**/*.spec.(ts|js){,x}'],
-      },
-      {
-        displayName: 'server',
-        testEnvironment: require.resolve('jest-environment-yoshi-bootstrap'),
-        testMatch: ['<rootDir>/test/server/**/*.spec.(ts|js){,x}'],
-        setupFiles: [
-          require.resolve(
-            'jest-environment-yoshi-bootstrap/environment-setup.js',
-          ),
-        ],
+        testMatch: [`<rootDir>/${globs.unitTests}`],
       },
       {
         displayName: 'e2e',
         testEnvironment: require.resolve('jest-environment-yoshi-puppeteer'),
-        testMatch: [`<rootDir>/${globs.puppeteer}`],
+        testMatch: [`<rootDir>/${globs.e2eTests}`],
         setupFiles: [
           require.resolve(
             'jest-environment-yoshi-bootstrap/environment-setup.js',
@@ -47,7 +37,7 @@ module.exports = {
       })
       .map(project => {
         const [setupTestsPath] = globby.sync(
-          `test/setup.${project.displayName}.(ts|js){,x}`,
+          `__tests__/${project.displayName}-setup.(ts|js){,x}`,
         );
 
         const setupTestsFile =
