+++ packages/jest-yoshi-preset/package.json
@@ -23,6 +23,7 @@
     "jest-environment-yoshi-bootstrap": "4.1.0-alpha.1",
     "jest-environment-yoshi-puppeteer": "4.1.0-alpha.1",
     "ts-jest": "^23.0.1",
+    "yoshi-config": "4.1.0-alpha.1",
     "yoshi-helpers": "4.1.0-alpha.1"
   },
   "devDependencies": {
