+++ packages/yoshi-config/globs.js
@@ -16,8 +16,8 @@ module.exports = {
   statics,
   babel: [path.join(base, '**', '*.js{,x}'), 'index.js'],
   specs: `${base}/**/*.+(spec|it).+(js|ts){,x}`,
-  protractor: `${test}/**/*.e2e.+(js|ts)`,
-  puppeteer: `${test}/e2e/**/*.spec.+(ts|js){,x}`,
+  e2eTests: '**/*.e2e.+(js|ts){,x}',
+  unitTests: '**/*.spec.+(ts|js){,x}',
   testFilesWatch: [path.join(base, '**', '*.(ts|js){,x}'), 'index.js'],
   singleModule: {
     clientDist: statics,
